FIX: resolved different viewport related bugs: 1. changed viewport dimensions in struct from int to double 2. corrected an inaccuracy in calling the function to add multiple vectors 3. Added cast to double when performing calculations with img dimensions. + some minor refactoring and addition of debug statements.

+++ includes/minirt.h
@@ -23,8 +23,8 @@ typedef struct	s_pxl {
 
 typedef struct s_viewp
 {
-	int		height;
-	int		width;
+	double	height;
+	double	width;
 	t_point	upperleft;
 	t_vec	uvec_w;
 	t_vec	uvec_h;
